fix cannot fetch bug

+++ backend/utilities.py
@@ -4,9 +4,7 @@ import os
 import csv
 import path
 
-img = cv.imread("../src/assets/TestData/lmTst_0.jpg")
-
-# resize
+# img = cv.imread("../src/assets/TestData/lmTst_0.jpg")
 
 
 def resize_image(image, scale_percent):
@@ -33,8 +31,6 @@ def image_draw_dot(image, coord):
 
     return image_result
 
-# rotate by degree around center
-
 
 def rotate_image(image, angle):
     (height, width) = image.shape[:2]
@@ -42,8 +38,6 @@ def rotate_image(image, angle):
     M = cv.getRotationMatrix2D((cX, cY), angle, 1.0)
     rotated = cv.warpAffine(image, M, (width, height))
     return rotated
-
-# translate
 
 
 def translate_image(image, hor, ver):
@@ -53,6 +47,8 @@ def translate_image(image, hor, ver):
     M = np.float32([[1, 0, hor], [0, 1, ver]])
     shifted = cv.warpAffine(image, M, (image.shape[1], image.shape[0]))
     return shifted
+
+# Create mock data
 
 
 def transform_mock(image_name, max_angle, max_x, max_y, destination):
@@ -69,49 +65,36 @@ def transform_mock(image_name, max_angle, max_x, max_y, destination):
                 f'{destination}/{image_name}_{i}_{j}_{0}.jpg', img_trans)
 
 
-img_name = "lmTst_0"
-dest = path.cur_image_path
+# img_name = "lmTst_0"
+# dest = path.cur_image_path
 # transform_mock(img_name, 10, 15, 15, dest)
 
 
 def write_debug(data, debug_file):
     for i in range(int(len(data))):
         data[i] = str(data[i])
-    # data = ['Afghanistan', 652090, 'AF', 'AFG']
-
-    # check if header already exist
 
     exist = False
+
+    # Check if debug_file exists
     if (os.path.exists(debug_file)):
+
+        # Check if data exists -> Skip writing that row
         with open(debug_file, 'r', encoding='UTF8') as read_file:
             csvreader = csv.reader(read_file, delimiter=',')
             for line in csvreader:
-                # # print(line)
-                # print(type(data))
                 if (line == data):
                     exist = True
 
         if not exist:
             with open(debug_file, 'a', newline='', encoding='UTF8') as f:
                 writer = csv.writer(f)
-
-                # write the header
-                # writer.writerow(header)
-
-                # write the data
                 writer.writerow(data)
-                # close file
                 f.close()
     else:
         with open(debug_file, 'a', newline='', encoding='UTF8') as f:
             writer = csv.writer(f)
-
-            # write the header
-            # writer.writerow(header)
-
-            # write the data
             writer.writerow(data)
-            # close file
             f.close()
 
 
@@ -128,8 +111,9 @@ def clear_file(file):
 # clear_file("debug/debug_log_Mock.csv")
 
 def remove_file(directory):
-    for f in os.listdir(directory):
-        os.remove(os.path.join(directory, f))
+    if (os.path.exists(directory)):
+        for f in os.listdir(directory):
+            os.remove(os.path.join(directory, f))
 
 
 # remove_file(f'{path.calibrated_path_mock}/Mock')
